Enable example doxmlparser possibilities for cmake build
The doxmlparser example program  could be build by the build system before the use of cmake. In cmake it was available but "disabled" by means of an empty CMakeLists.txt in the doxmlparser main directory.
(Windows is not fond of symbolic links, so removing link between src/doxmlintf.h and include/doxmlintf.h)